refactor: 페이지별로 리팩토링
