+++ static/js/page/renderer.js
@@ -12,40 +12,40 @@ import {
     SubMenu,
 } from "../views/views.js";
 
-const render = (html) => {
-    $("#app").innerHTML = html;
-};
+// const render = (html) => {
+//     $("#app").innerHTML = html;
+// };
 
-const renderSubMenu = () => {
-    $(".sub-menu").innerHTML = new SubMenu({
-        menus: datas.subMenu.webtoon.menuNames,
-    }).getHtml();
-};
+// const renderSubMenu = () => {
+//     $(".sub-menu").innerHTML = new SubMenu({
+//         menus: datas.subMenu.webtoon.menuNames,
+//     }).getHtml();
+// };
 
-const renderHome = () => {
-    const category = new Category({ categories: datas.homeData.category });
-    const slideBanner = new SlideBanner();
-    let html = "";
-    html += category.getHtml();
-    html += slideBanner.getHtml();
+// const renderHome = () => {
+//     const category = new Category({ categories: datas.homeData.category });
+//     const slideBanner = new SlideBanner();
+//     let html = "";
+//     html += category.getHtml();
+//     html += slideBanner.getHtml();
 
-    const renderJSONData = (path) => {
-        getJSON(path)
-            .then((data) => {
-                html += new GenreBest({
-                    genre: "로맨스",
-                    genreItem: data,
-                }).getHtml();
+//     const renderJSONData = (path) => {
+//         getJSON(path)
+//             .then((data) => {
+//                 html += new GenreBest({
+//                     genre: "로맨스",
+//                     genreItem: data,
+//                 }).getHtml();
 
-                render(html);
-            })
-            .catch(render(html));
-    };
+//                 render(html);
+//             })
+//             .catch(render(html));
+//     };
 
-    renderSubMenu();
-    renderJSONData("genre/romanceTop");
-    renderCarousel(datas.carouselImgs["Home"]);
-};
+//     renderSubMenu();
+//     renderJSONData("genre/romanceTop");
+//     renderCarousel(datas.carouselImgs["Home"]);
+// };
 
 const renderWeekday = (today = new Date().getDay()) => {
     const days = ["일", "월", "화", "수", "목", "금", "토"];
@@ -60,4 +60,7 @@ const renderWeekday = (today = new Date().getDay()) => {
     render(html);
 };
 
-export { render, renderHome, renderWeekday };
+export {
+    // render, renderHome,
+    renderWeekday,
+};
